inline_bot/pyproject.toml
[project]
name = "inline_bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]
